LMS/lms_app/backend/requirements.txt
Flask
Flask-PyMongo
flask-jwt-extended
flask-cors
python-dotenv
